Motor Control

+++ Tasks/Comm.h
@@ -89,7 +89,7 @@ typedef struct {
 	int32_t vac_pressure;
 	int16_t temperature;
 	int16_t targetTemperature;
-	int16_t x_encoder_pos;
+	float x_encoder_pos;
 	uint8_t motor_state;
 	TriggerState_t trigger_state;
 	char fpath[32];
